+++ src/Management/WhatsNewProvider.php
@@ -1,0 +1,22 @@
+<?php
+/*
+ * (c) 2026: 975L <contact@975l.com>
+ * (c) 2026: Laurent Marquet <laurent.marquet@laposte.net>
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+namespace c975L\ContactFormBundle\Management;
+
+use c975L\ConfigBundle\Management\WhatsNewJsonReader;
+use c975L\ConfigBundle\Management\WhatsNewProviderInterface;
+
+class WhatsNewProvider implements WhatsNewProviderInterface
+{
+    private const BUNDLE_NAME = 'ContactFormBundle';
+
+    public function getEntries(): array
+    {
+        return WhatsNewJsonReader::read(\dirname(__DIR__, 2) . '/config/whatsnew.json', self::BUNDLE_NAME);
+    }
+}